unbreak compilation
svn path=/trunk/kdeedu/kig/; revision=253120

+++ misc/kigpainter.cpp
@@ -752,7 +752,7 @@ void KigPainter::drawAngle( const Coordinate& cpoint, const double dstartangle,
   QPoint end( static_cast<int>( point.x() + radius * cos( dstartangle + dangle ) ),
               static_cast<int>( point.y() - radius * sin( dstartangle + dangle ) ) );
   QPoint vect = (end - point);
-  double vectlen = sqrt( vect.x() * vect.x() + vect.y() * vect.y() );
+  double vectlen = sqrt( float( vect.x() * vect.x() + vect.y() * vect.y() ) );
   QPoint orthvect( -vect.y(), vect.x() );
   vect = vect * 6 / vectlen;
   orthvect = orthvect * 6 / vectlen;
